requirements.txt
Django==4.2.1
Pillow==9.5.0
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3
